bigger cleanup

+++ DeepReader/Classes/StringExtensions.cs
@@ -26,7 +26,7 @@ public static class StringExtensions
 
     public static int GetHexVal(char hex)
     {
-        int val = (int)hex;
+        int val = hex;
         //For uppercase A-F letters:
         //return val - (val < 58 ? 48 : 55);
         //For lowercase a-f letters:
